scripts/requirements.txt
nsepy==0.8
mysql-connector-python==8.0.21
Flask==1.1.2
Flask-Cors==3.0.9
cython
statsmodels==0.12.0
matplotlib==3.3.1
